Refactoring and improvements

+++ BankingControlPanel/Controllers/ClientsController.cs
@@ -1,6 +1,7 @@
 ﻿using BankingControlPanel.Contracts;
 using BankingControlPanel.DTOs;
 using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BankingControlPanel.Controllers
@@ -17,43 +18,56 @@ namespace BankingControlPanel.Controllers
             _service = service;
             _validator = validator;
         }
+
         [HttpPost]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> AddClient(ClientDto clientDto)
         {
-
             var validationError = await ValidateChecksForAdd(clientDto);
             if (validationError != null)
             {
                 return validationError;
             }
 
-            await _service.AddClient(clientDto);
-            return CreatedAtAction(nameof(GetClientById), clientDto);
+            var client = await _service.AddClient(clientDto);
+            return Ok(client);
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteClient(int id)
         {
+            var client = await _service.GetClientById(id);
+            if (client == null)
+            {
+                return NotFound("No Record Found");
+            }
             await _service.DeleteClient(id);
             return NoContent();
         }
+
         [HttpGet("{id}")]
+        [Authorize(Roles = "Admin, User")]
         public async Task<IActionResult> GetClientById(int id)
         {
             var client = await _service.GetClientById(id);
             if (client == null)
             {
-                return NotFound();
+                return NotFound("No record found with this client id");
             }
             return Ok(client);
         }
+
         [HttpGet]
+        [Authorize(Roles = "Admin, User")]
         public async Task<IActionResult> GetClients([FromQuery] ClientParameters parameters)
         {
             var clients = await _service.GetClients(parameters);
             return Ok(clients);
         }
+
         [HttpGet("filtered")]
+        [Authorize(Roles = "Admin, User")]
         public async Task<IActionResult> GetFilteredClients([FromQuery] ClientParameters parameters)
         {
             if (string.IsNullOrWhiteSpace(parameters.FirstName) &&
@@ -66,13 +80,11 @@ namespace BankingControlPanel.Controllers
             var lastSearchParameters = _service.GetLastSearchParameters();
             return Ok(new { Clients = clients, LastSearchParameters = lastSearchParameters });
         }
+
         [HttpGet("sorted")]
+        [Authorize(Roles = "Admin, User")]
         public async Task<IActionResult> GetSortedClients([FromQuery] ClientParameters parameters)
         {
-            if (string.IsNullOrWhiteSpace(parameters.SortBy) || string.IsNullOrWhiteSpace(parameters.SortOrder))
-            {
-                return BadRequest("SortBy and SortOrder must be provided for sorting.");
-            }
             if (string.IsNullOrWhiteSpace(parameters.SortBy) || string.IsNullOrWhiteSpace(parameters.SortOrder))
             {
                 return BadRequest("SortBy and SortOrder must be provided for sorting.");
@@ -82,10 +94,12 @@ namespace BankingControlPanel.Controllers
             var lastSearchParameters = _service.GetLastSearchParameters();
             return Ok(new { Clients = clients, LastSearchParameters = lastSearchParameters });
         }
+
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateClient(int id, ClientDto clientDto)
+        [Authorize(Roles = "Admin")]
+        public async Task<IActionResult> UpdateClient(int id, [FromBody] ClientDto clientDto)
         {
-            var validationError = await ValidateChecksForUpdate(id,clientDto);
+            var validationError = await ValidateChecksForUpdate(id, clientDto);
             if (validationError != null)
             {
                 return validationError;
@@ -95,6 +109,7 @@ namespace BankingControlPanel.Controllers
 
             return NoContent();
         }
+
         private async Task<IActionResult> ValidateChecksForAdd(ClientDto clientDto)
         {
             var validationResult = await _validator.ValidateAsync(clientDto);
@@ -115,14 +130,26 @@ namespace BankingControlPanel.Controllers
                 return BadRequest(new { message = "Personal ID already exists" });
             }
 
-            if (clientDto.Accounts != null && clientDto.Accounts.Any(account => account.Balance < 0))
+            if (clientDto.Accounts != null)
             {
-                return BadRequest(new { message = "Balance cannot be negative" });
-            }
+                foreach (var account in clientDto.Accounts)
+                {
+                    if (account.Balance < 0)
+                    {
+                        return BadRequest(new { message = "Balance cannot be negative" });
+                    }
 
+                    var existingAccountNumber = await _service.AccountNumberExists(account.AccountNumber);
+                    if (existingAccountNumber)
+                    {
+                        return BadRequest(new { message = $"Account number {account.AccountNumber} already exists" });
+                    }
+                }
+            }
             return null;
         }
-        private async Task<IActionResult> ValidateChecksForUpdate(int id,ClientDto clientDto)
+
+        private async Task<IActionResult> ValidateChecksForUpdate(int id, ClientDto clientDto)
         {
             var validationResult = await _validator.ValidateAsync(clientDto);
             if (!validationResult.IsValid)
@@ -145,6 +172,7 @@ namespace BankingControlPanel.Controllers
                 return BadRequest(new { message = "Balance cannot be negative" });
             }
 
-            return null; }
+            return null;
+        }
     }
 }
